# Enh/rewrite interface