* add uniform `string_repr(w::AbstractWord, context)` where `context` could be either `Alphabet` or `Ordering` 
* cleaned-up `rewrite_from_left`
* simplified lots of signatures from `automata.jl`
* integrated `Automaton` and its rewriting into it
* some performance tunning, notably: added `_past_states` field to `Automaton` to be re-used in `rewrite_from_left`

```julia
using Test, BenchmarkTools, Random
using KnuthBendix
include("test/rws_examples.jl")
```
before:
```julia
julia> let R = RWS_Example_237_abaB(4), seed=1
           rws = KnuthBendix.knuthbendix2(R)

           n = length(KnuthBendix.alphabet(KnuthBendix.ordering(rws)))
           Random.seed!(seed)
           w = Word(rand(1:n, 2000))

           @btime KnuthBendix.rewrite_from_left($w, $rws)
           @time ia = KnuthBendix.makeindexautomaton(rws, KnuthBendix.alphabet(KnuthBendix.ordering(rws)))
           @btime KnuthBendix.index_rewrite($w, $ia)
       end;
  1.453 ms (5 allocations: 8.30 KiB)
  0.000742 seconds (4.76 k allocations: 218.672 KiB)
  242.169 μs (4430 allocations: 229.52 KiB)
```

after
```julia
julia> let R = RWS_Example_237_abaB(4), seed=1
           rws = KnuthBendix.knuthbendix2(R)

           n = length(KnuthBendix.alphabet(KnuthBendix.ordering(rws)))
           Random.seed!(seed)
           w = Word(rand(1:n, 2000))

           @btime KnuthBendix.rewrite_from_left($w, $rws)
           @time ia = KnuthBendix.makeindexautomaton(rws)
           @btime KnuthBendix.rewrite_from_left($w, $ia)
       end;
  1.170 ms (5 allocations: 8.30 KiB)
  0.000483 seconds (4.76 k allocations: 218.750 KiB)
  98.662 μs (5 allocations: 8.30 KiB)
```